+++ src/lib/string.ts
@@ -1,0 +1,34 @@
+/**
+ * Indicates whether a specified string is null/undefined or empty
+ * @param value The string to test
+ * @returns true if the value parameter is null/undefined or empty
+ */
+export function isNullOrEmpty(value?: string): boolean {
+  if (!value || typeof value !== "string") {
+    return true;
+  }
+
+  return value.length === 0;
+}
+
+/**
+ * Indicates whether a specified string is null/undefined, empty, or consists only of white-space characters
+ * @param value The string to test
+ * @returns true if the value parameter is null/undefined, Empty, or if value consists exclusively of white-space characters
+ */
+export function isNullOrWhitespace(value?: string): boolean {
+  return isNullOrEmpty(value) || (value ?? "").trim().length === 0;
+}
+
+/**
+ * Capitalize the string
+ * @param value The string to capitalize
+ * @returns true The capitalized string
+ */
+export function capitalize(value?: string): string | undefined {
+  if (!value || isNullOrWhitespace(value)) {
+    return value;
+  }
+
+  return value.charAt(0).toUpperCase() + value.slice(1);
+}